+++ R/estimates.R
@@ -1,63 +1,66 @@
-#' Estimates
-#'
-#' Get the estimates for an MCMC object.
-#'
-#' @param object An MCMC object.
-#' @param fun The function to use.
-#' @param as_df A flag indicating whether to return the estimates as a data frame versus a named list.
+#' @export
+universals::estimates
+
+#' @inherit universals::estimates
+#' @inheritParams params
+#' @return A named list or data frame.
+#' @export
+estimates.mcarray <- function(x, fun = median, as_df = FALSE, ...) {
+  if(!as_df)
+    return(estimates(as.mcmcarray(x), fun = fun, as_df = FALSE))
+  estimates(as.mcmc.list(x), fun = fun, as_df = TRUE)
+}
+
+#' @inherit universals::estimates
+#' @inheritParams params
+#' @return A named list or data frame.
+#' @export
+estimates.mcmc <- function(x, fun = median, as_df = FALSE, ...) {
+  if(!as_df)
+    return(estimates(as.mcmcr(x), fun = fun, as_df = FALSE))
+
+  term <- as.term(x)
+  x <- t(x)
+  x <- .estimates(x, fun = fun)
+  tibble(term = term, estimate = x)
+}
+
+#' @inherit universals::estimates
+#' @inheritParams params
+#' @return A named list or data frame.
+#' @export
+estimates.mcmc.list <- function(x, fun = median, as_df = FALSE, ...) {
+  if(!as_df)
+    return(estimates(as.mcmcr(x), fun = fun, as_df = FALSE))
+
+  x <- as.mcmc(collapse_chains(x))
+  estimates(x, fun = fun, as_df = TRUE)
+}
+
+#' @inherit universals::estimates
+#' @inheritParams params
 #' @param ... Unused.
 #' @return A named list or data frame.
 #' @export
-#' @examples
-#' estimates(mcmcr_example)
-#' estimates(mcmcr_example, as_df = TRUE)
-estimates <- function(object, ...) {
-  UseMethod("estimates")
-}
-
-#' @describeIn estimates Estimates mcarray
-#' @export
-estimates.mcarray <- function(object, fun = stats::median, as_df = FALSE, ...) {
-  if(!as_df)
-    return(estimates(as.mcmcarray(object), fun = fun, as_df = FALSE))
-  estimates(as.mcmc.list(object), fun = fun, as_df = TRUE)
-}
-
-#' @export
-estimates.mcmc <- function(object, fun = stats::median, as_df = FALSE, ...) {
-  if(!as_df)
-    return(estimates(as.mcmcr(object), fun = fun, as_df = FALSE))
-
-  term <- as.term(object)
-  object <- t(object)
-  object <- .estimates(object, fun = fun)
-  tibble(term = term, estimate = object)
-}
-
-#' @export
-estimates.mcmc.list <- function(object, fun = stats::median, as_df = FALSE, ...) {
-  if(!as_df)
-    return(estimates(as.mcmcr(object), fun = fun, as_df = FALSE))
-
-  object <- as.mcmc(collapse_chains(object))
-  estimates(object, fun = fun, as_df = TRUE)
-}
-
-#' @export
-estimates.mcmcarray <- function(object, fun = stats::median, as_df = FALSE, ...) {
+estimates.mcmcarray <- function(x, fun = median, as_df = FALSE, ...) {
   chk_function(fun)
   chk_flag(as_df)
 
   if(as_df)
-    return(estimates(as.mcmc.list(object), fun = fun, as_df = TRUE))
+    return(estimates(as.mcmc.list(x), fun = fun, as_df = TRUE))
 
-  object <- collapse_chains(object)
-  apply(object, 3:ndims(object), FUN = .estimates, fun = fun)
+  x <- collapse_chains(x)
+  apply(x, 3:ndims(x), FUN = .estimates, fun = fun)
 }
 
+#' @inherit universals::estimates
+#' @inheritParams params
+#' @return A named list or data frame.
 #' @export
-estimates.mcmcr <- function(object, fun = stats::median, as_df = FALSE, ...) {
+#' @examples
+#' estimates(mcmcr_example)
+estimates.mcmcr <- function(x, fun = median, as_df = FALSE, ...) {
   if(as_df)
-    return(estimates(as.mcmc.list(object), fun = fun, as_df = TRUE))
-  lapply(object, estimates, fun = fun, as_df = FALSE, ...)
+    return(estimates(as.mcmc.list(x), fun = fun, as_df = TRUE))
+  lapply(x, estimates, fun = fun, as_df = FALSE, ...)
 }