+++ R/export.R
@@ -1,0 +1,2 @@
+#' @export
+stats::median