+++ R/namespace.R
@@ -1,5 +1,5 @@
 #' @import chk universals term nlist extras
 #' @importFrom coda as.mcmc.list as.mcmc thin mcmc.list
 #' @importFrom graphics plot
-#' @importFrom stats coef terms
+#' @importFrom stats coef terms median
 NULL