+++ R/set-pars.R
@@ -3,34 +3,33 @@ universals::set_pars
 
 #' @inherit universals::set_pars
 #' @export
-set_pars.mcmc <- function(x, value) {
-  term <- set_pars(as.term(x), value)
+set_pars.mcmc <- function(x, value, ...) {
+  term <- set_pars(as.term(x), value, ...)
   colnames(x) <- as.character(term)
   x
 }
 
 #' @inherit universals::set_pars
 #' @export
-set_pars.mcmc.list <- function(x, value) {
-  x <- lapply(x, set_pars, value = value)
+set_pars.mcmc.list <- function(x, value, ...) {
+  x <- lapply(x, set_pars, value = value, ...)
   class(x) <- "mcmc.list"
   x
 }
 
 #' @inherit universals::set_pars
 #' @export
-set_pars.mcmcr <- function(x, value) {
-  chk_s3_class(value, "character")
+set_pars.mcmcr <- function(x, value, ...) {
+  chk_pars(value)
   chk_identical(length(value), length(x))
-  chk_unique(value)
   names(x) <- value
   x
 }
 
 #' @inherit universals::set_pars
 #' @export
-set_pars.mcmcrs <- function(x, value) {
-  x <- lapply(x, `pars<-`, value)
+set_pars.mcmcrs <- function(x, value, ...) {
+  x <- lapply(x, `pars<-`, value, ...)
   class(x) <- "mcmcrs"
   x
 }